movie_theater_booking/requirements.txt
asgiref==3.7.2
attrs==25.4.0
behave==1.3.3
certifi==2026.2.25
colorama==0.4.6
cucumber-expressions==19.0.0
cucumber-tag-expressions==9.1.0
Django==4.2.11
django-behave==0.1.6
djangorestframework==3.16.1
gunicorn==25.1.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
outcome==1.3.0.post0
packaging==26.0
parse==1.21.1
parse_type==0.6.6
pluggy==1.6.0
Pygments==2.19.2
PySocks==1.7.1
pytest==9.0.2
pytz==2024.1
selenium==4.41.0
setuptools==68.1.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
supervisor==4.2.5
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wheel==0.42.0
wsproto==1.3.2
